health: #184 moved test/facade to test/content and rearranged by topic (instead of by type)
