--- packages/core/src/test/facade/data/index.ts
@@ -1,3 +1,0 @@
-export * from "./generate-tree-nodes";
-export * from "./shape-repository";
-export * from "./tree-node-repository";
