--- packages/core/src/test/facade/model/index.ts
@@ -1,8 +1,0 @@
-export * from "./author.model";
-export * from "./data-entry.model";
-export * from "./metadata.model";
-export * from "./shape.model";
-export * from "./tree-node-level.model";
-export * from "./tree-node-parents.model";
-export * from "./tree-node.model";
-export * from "./user.model";
